requirements.txt
colorama~=0.4.6
pyyaml~=6.0.2
pydantic~=2.9.2
langchain~=0.3.1
langchain-community~=0.3.1

